Create billPulseCounter
The provided Python script utilizes the RPi.GPIO library to count pulses received on a specific GPIO pin of the Raspberry Pi. It sets up an event detection on the rising edge of the input signal and increments a pulse counter each time a valid pulse is detected. To mitigate potential noise or false triggers, a debounce time is implemented to ignore subsequent pulses within a specified interval.